video-rag-gemini/requirements.txt
# Streamlit Video RAG Demo Requirements
streamlit>=1.28.0
google-generativeai>=0.3.0
requests>=2.31.0
Pillow>=10.0.0

